+++ code/ylikuutio/ontology/loaders/heightmap_loader.hpp
@@ -25,8 +25,8 @@ namespace ontology
 
     bool load_SRTM_world(
             std::string image_path,
-            double latitude,
-            double longitude,
+            float latitude,
+            float longitude,
             std::vector<glm::vec3>& out_vertices,
             std::vector<glm::vec2>& out_UVs,
             std::vector<glm::vec3>& out_normals,
